requirements.txt
ultralytics
opencv-python
numpy
matplotlib
seaborn
scikit-learn
PyYAML
Pillow
torch
